% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{add_rdg_wri_subscores}
\alias{add_rdg_wri_subscores}
\title{Add reading and writing subscores to the items data file}
\usage{
add_rdg_wri_subscores(items)
}
\arguments{
\item{items}{items table from \code{get_items_table()}}
}
\value{
a data frame of item scores with rdg and writing subscores added
}
\description{
Add reading and writing subscores to the items data file
}
\keyword{Internal}
