% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_db}
\alias{current_db}
\title{Used in case the db is not passed to a function
Determines the current db based on the system date
Creates a string with the name of the db}
\usage{
current_db()
}
\description{
Used in case the db is not passed to a function
Determines the current db based on the system date
Creates a string with the name of the db
}
\keyword{internal}
