% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbprocess-package.R
\docType{package}
\name{dbprocess-package}
\alias{dbprocess}
\alias{dbprocess-package}
\title{dbprocess: Processes data from the ORExt database in preparation for analysis}
\description{
This package is designed to be used in conjunction with the orextdb package. The orextdb package retrieves data from the live ORExt database, through a secure web API, and this package processes the data. In particular, you should not need to ever retrieve specific tables from the live DB, becuase this package will pull the table needed and combine them in a way that they are ready for analysis.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

}
\keyword{internal}
