% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors-helpers.R
\name{get_items_table}
\alias{get_items_table}
\title{internal function to format the items table from other tables. see \code{get_items()} in processors.R}
\usage{
get_items_table(
  submissions_table,
  student_table,
  exam_table,
  tasks_table,
  answer_table,
  items_table
)
}
\description{
internal function to format the items table from other tables. see \code{get_items()} in processors.R
}
\keyword{internal}
