% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_pattern_data}
\alias{get_pattern_data}
\title{Create patterned synthetic data for any (or all) tests related to all
possible raw scores}
\usage{
get_pattern_data(name = NULL, grade = NULL, content = NULL)
}
\arguments{
\item{name}{The name of the test to download (e.g., Science_G5, ELA_G11). If
used, subsequent arguments to \code{grade} and \code{content} are ignored.}
}
\value{
Similar to \code{get_test_json}, if \code{name} or both
\code{grade} and \code{content} are supplied, the patterned data
for just that test is returned. Otherwise, patterned data for
all tests is returned.The patterned data is synthetic and created
such that all possible raw scores are generated. This function
is primarily used to create the raw to scale score mapping.
}
\description{
Create patterned synthetic data for any (or all) tests related to all
possible raw scores
}
