% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_test_json}
\alias{get_test_json}
\title{Get the JSON data for any (or all) tests}
\usage{
get_test_json(name = NULL, grade = NULL, content = NULL)
}
\arguments{
\item{name}{The name of the test to download (e.g., Science_G5, ELA_G11). If
used, subsequent arguments to \code{grade} and \code{content} are ignored.}

\item{grade}{Optional argument to return only data from a specific grade.
Default is \code{NULL}, in which case all grades are returned in a list,
where each element of the list is a data frame for the specific test form.}

\item{content}{Optional argument to return only data from a specific content
area. Default is \code{NULL}, in which case all content areas are returned
in a list, where each element of the list is a data frame for the specific
test form. Should be one of \code{"ELA"}, \code{"Math"}, or
\code{"Science"}, although case does not matter (e.g., \code{"ela"}
will also work).}
}
\value{
If \code{name} or \code{grade} and \code{content} are supplied, the
json data for just that test is returned. If all are \code{NULL}, then
a list with the json data for all tests are returned.
}
\description{
Get the JSON data for any (or all) tests
}
\keyword{internal}
