% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{pull_item_ids}
\alias{pull_item_ids}
\title{Pull the item IDS from the json data}
\usage{
pull_item_ids(json)
}
\arguments{
\item{json}{A single JSON file from \code{get_test_json()}
(internal dbprocess function)}
}
\value{
a character vector of item ids
}
\description{
Pull the item IDS from the json data
}
\keyword{internal}
